requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
openai>=1.68.2
python-dotenv>=1.0.1
python-multipart>=0.0.9
Pillow>=10.2.0
sqlalchemy
aiomysql
pydantic[email]
python-jose[cryptography]
bcrypt
pymysql
